feature: update Button component in Storybook
